moved passthrough controller to storage subfolder
